# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/communities": {
      "get": {
        "description": "Lista comunidades ativas com paginação HATEOAS e filtros",
        "operationId": "CommunityController_getAll",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Quantidade de registros",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Quantidade de registros para pular",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrar por status ativo/inativo",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Buscar por nome ou descrição",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de comunidades retornada com sucesso com links de navegação"
          }
        },
        "summary": "Listar todas as comunidades",
        "tags": [
          "communities"
        ]
      },
      "post": {
        "description": "Cria uma nova comunidade para o usuário autenticado",
        "operationId": "CommunityController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comunidade criada com sucesso"
          },
          "403": {
            "description": "Usuário não tem permissão para criar comunidade"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar nova comunidade",
        "tags": [
          "communities"
        ]
      }
    },
    "/api/v1/communities/me": {
      "get": {
        "description": "Retorna os dados da comunidade do usuário autenticado",
        "operationId": "CommunityController_getMyCommunity",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dados da comunidade retornados com sucesso"
          },
          "403": {
            "description": "Usuário não tem perfil de comunidade"
          },
          "404": {
            "description": "Comunidade não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter minha comunidade",
        "tags": [
          "communities"
        ]
      }
    },
    "/api/v1/communities/{id}": {
      "get": {
        "operationId": "CommunityController_getByID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da comunidade"
          },
          "404": {
            "description": "Comunidade não encontrada"
          }
        },
        "summary": "Buscar comunidade por ID",
        "tags": [
          "communities"
        ]
      },
      "patch": {
        "description": "Atualiza os dados da comunidade. Apenas o dono pode atualizar",
        "operationId": "CommunityController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommunityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comunidade atualizada com sucesso"
          },
          "403": {
            "description": "Sem permissão para atualizar esta comunidade"
          },
          "404": {
            "description": "Comunidade não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar comunidade",
        "tags": [
          "communities"
        ]
      },
      "delete": {
        "description": "Remove uma comunidade. Apenas o dono pode deletar",
        "operationId": "CommunityController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comunidade deletada com sucesso"
          },
          "403": {
            "description": "Sem permissão para deletar esta comunidade"
          },
          "404": {
            "description": "Comunidade não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletar comunidade",
        "tags": [
          "communities"
        ]
      }
    },
    "/api/v1/auth/signin": {
      "post": {
        "description": "Realiza login com email e senha. Retorna dados do usuário e seus papéis",
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        },
        "summary": "Autenticar usuário",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/signout": {
      "post": {
        "description": "Realiza logout do usuário autenticado",
        "operationId": "AuthController_signOut",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout realizado com sucesso"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Encerrar sessão",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/password/reset/token": {
      "post": {
        "description": "Envia um email com link de recuperação se o usuário existir",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitação processada"
          }
        },
        "summary": "Solicitar recuperação de senha",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "description": "Altera a senha do usuário usando o token recebido por email",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso"
          },
          "409": {
            "description": "Token inválido ou expirado"
          }
        },
        "summary": "Redefinir senha",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Retorna dados completos do usuário autenticado incluindo papéis e permissões",
        "operationId": "AuthController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dados do usuário retornados com sucesso"
          },
          "401": {
            "description": "Não autenticado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter perfil do usuário logado",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/admin/users": {
      "post": {
        "description": "Permite que administradores criem novos usuários com papéis específicos",
        "operationId": "AuthController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          },
          "403": {
            "description": "Acesso negado - apenas administradores"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[ADMIN] Criar novo usuário",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/admin/communities": {
      "post": {
        "description": "Permite que administradores criem comunidades",
        "operationId": "AuthController_createCommunity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunitySignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comunidade criada com sucesso"
          },
          "403": {
            "description": "Acesso negado - apenas administradores"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[ADMIN] Criar nova comunidade",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/bootstrap/admin": {
      "post": {
        "description": "Rota de inicialização para criar o primeiro admin. Bloqueada após primeiro uso",
        "operationId": "AuthController_bootstrapAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Administrador criado com sucesso"
          },
          "409": {
            "description": "Já existe um administrador no sistema"
          }
        },
        "summary": "Criar primeiro administrador do sistema",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/signup/community": {
      "post": {
        "description": "Permite que comunidades se registrem diretamente na plataforma",
        "operationId": "AuthController_signUpCommunity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommunitySignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comunidade cadastrada com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "409": {
            "description": "Email já cadastrado"
          }
        },
        "summary": "Cadastro público de comunidade",
        "tags": [
          "authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "EventDev API",
    "description": "API RESTful para gerenciamento de eventos e comunidades com autorização baseada em papéis e permissões",
    "version": "1.0",
    "contact": {
      "name": "EventDev Team",
      "url": "https://eventdev.org",
      "email": "contact@eventdev.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:5122",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Token de autenticação JWT do SuperTokens",
        "in": "header"
      }
    },
    "schemas": {
      "CreateCommunityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tech Community Brasil",
            "description": "Nome da comunidade"
          },
          "description": {
            "type": "string",
            "example": "Comunidade focada em tecnologia"
          },
          "logoUrl": {
            "type": "string",
            "example": "https://example.com/logo.png"
          },
          "phoneNumber": {
            "type": "string",
            "example": "11987654321"
          },
          "instagramLink": {
            "type": "string",
            "example": "https://instagram.com/techcommunity"
          },
          "linkedinLink": {
            "type": "string",
            "example": "https://linkedin.com/company/techcommunity"
          },
          "websiteLink": {
            "type": "string",
            "example": "https://techcommunity.dev"
          },
          "githubLink": {
            "type": "string",
            "example": "https://github.com/techcommunity"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "default": true
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCommunityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tech Community Brasil",
            "description": "Nome da comunidade"
          },
          "description": {
            "type": "string",
            "example": "Comunidade focada em tecnologia"
          },
          "logoUrl": {
            "type": "string",
            "example": "https://example.com/logo.png"
          },
          "phoneNumber": {
            "type": "string",
            "example": "11987654321"
          },
          "instagramLink": {
            "type": "string",
            "example": "https://instagram.com/techcommunity"
          },
          "linkedinLink": {
            "type": "string",
            "example": "https://linkedin.com/company/techcommunity"
          },
          "websiteLink": {
            "type": "string",
            "example": "https://techcommunity.dev"
          },
          "githubLink": {
            "type": "string",
            "example": "https://github.com/techcommunity"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "default": true
          }
        }
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email do usuário"
          },
          "password": {
            "type": "string",
            "example": "Password@123",
            "description": "Senha do usuário"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email do usuário para recuperação de senha"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "token-xyz-123",
            "description": "Token de redefinição de senha recebido por email"
          },
          "password": {
            "type": "string",
            "example": "NewPassword123!",
            "description": "Nova senha do usuário"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "UserSignUpDto": {
        "type": "object",
        "properties": {}
      },
      "CommunitySignUpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "community@example.com",
            "description": "Email da comunidade"
          },
          "name": {
            "type": "string",
            "example": "Tech Community Brasil",
            "description": "Nome da comunidade"
          },
          "description": {
            "type": "string",
            "example": "Comunidade focada em tecnologia"
          },
          "logoUrl": {
            "type": "string",
            "example": "https://example.com/logo.png"
          },
          "phoneNumber": {
            "type": "string",
            "example": "11987654321"
          },
          "instagramLink": {
            "type": "string",
            "example": "https://instagram.com/techcommunity"
          },
          "linkedinLink": {
            "type": "string",
            "example": "https://linkedin.com/company/techcommunity"
          },
          "websiteLink": {
            "type": "string",
            "example": "https://techcommunity.dev"
          },
          "githubLink": {
            "type": "string",
            "example": "https://github.com/techcommunity"
          },
          "role": {
            "type": "string",
            "example": "community",
            "enum": [
              "user",
              "community"
            ],
            "description": "Tipo de conta"
          }
        },
        "required": [
          "email",
          "name",
          "role"
        ]
      },
...
    }
  }
}